+++ TypeContractor/Casing.cs
@@ -1,0 +1,24 @@
+﻿namespace TypeContractor;
+
+public enum Casing
+{
+	/// <summary>
+	/// PascalCase (default)
+	/// </summary>
+	Pascal,
+
+	/// <summary>
+	/// camelCase
+	/// </summary>
+	Camel,
+
+	/// <summary>
+	/// kebab-case
+	/// </summary>
+	Kebab,
+
+	/// <summary>
+	/// snake_case
+	/// </summary>
+	Snake,
+}